Move cache_name to uv-cache-key
