+++ crates/uv-cache-key/src/lib.rs
@@ -1,6 +1,6 @@
 pub use cache_key::{CacheKey, CacheKeyHasher};
 pub use canonical_url::{CanonicalUrl, RepositoryUrl};
-pub use digest::{cache_digest, hash_digest};
+pub use digest::{cache_digest, cache_name, hash_digest};
 
 mod cache_key;
 mod canonical_url;
